crawler/requirements.txt
redis
schedule
google-play-scraper
psycopg2-binary